Refactor settings tokens page to redirect to profile tokens; update app shell to reflect profile instead of settings; adjust stat card styles for consistency
